GoLang/Authentication/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
passlib[bcrypt]
pyjwt
pyotp
python-multipart
pydantic[email]
email-validator
httpx